+++ packages/log-structured/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+-   An issue when formatting a request log that has no headers.
+
 ## v1.0.3 - 2021-10-26
 
 ### Add